Moved computer guess to top of loop
Instead of generating a guess at the end of the proccess_feedback function and pushing it to the guess_list, I have a new guess generated right as needed.

Checking for loss due to lack of guesses was moved from the game_over? method into the game loop for guessing the code.